+++ apps/timeshift/apps.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig
+
+
+class TimeshiftConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "apps.timeshift"
+    verbose_name = "Timeshift (XC Catch-up)"
+    label = "timeshift"